feat(resume): moved resume schema to server only module
Updated import paths in basics.tsx, education.tsx, experiences.tsx,
interests.tsx, languages.tsx and skills.tsx. Added new schema test files
 for award, basic, certificates, education and email. Created new
 education, email, award, certificates and basic schemas.
