+++ apps/web/app/.server/schemas/resume/basics.ts
@@ -1,5 +1,6 @@
 // biome-ignore lint/nursery/noNamespaceImport: this is how we import from schema
 import * as S from '@effect/schema/Schema'
+
 import { Email } from './email.ts'
 import { Location } from './location.ts'
 import { Phone } from './phone.ts'
