+++ apps/web/app/.server/schemas/resume/certificates.ts
@@ -1,4 +1,5 @@
 import * as S from '@effect/schema/Schema'
+
 import { UrlString } from './url-string.ts'
 
 export const Certificate = S.struct({
