+++ apps/web/app/.server/schemas/resume/education.ts
@@ -1,5 +1,6 @@
 // biome-ignore lint/nursery/noNamespaceImport: this is how we import from schema
 import * as S from '@effect/schema/Schema'
+
 import { ISO8601Date } from './iso8601-date.ts'
 import { UrlString } from './url-string.ts'
