+++ apps/web/app/.server/schemas/resume/resume.spec.ts
@@ -1,8 +1,10 @@
 import { ParseError } from '@effect/schema/ParseResult'
 import * as Schema from '@effect/schema/Schema'
-import resumeAssetUrl from '@suddenly-giovanni/resume/resume.json?raw'
 import * as Either from 'effect/Either'
 import { describe, expect, it } from 'vitest'
+
+import resumeAssetUrl from '@suddenly-giovanni/resume/resume.json?raw'
+
 import type { Basics } from './basics.ts'
 import { type Resume as ResumeEncoded, Resume as ResumeSchema } from './resume.ts'
