+++ apps/web/app/.server/schemas/resume/skill.spec.ts
@@ -1,6 +1,7 @@
 // biome-ignore lint/nursery/noNamespaceImport: this is how we import from schema
 import * as S from '@effect/schema/Schema'
 import { describe, expect, test } from 'vitest'
+
 import { Skill } from './skill.ts'
 
 describe('Skill', () => {
