+++ apps/web/app/.server/schemas/resume/work.ts
@@ -1,5 +1,6 @@
 // biome-ignore lint/nursery/noNamespaceImport: this is how we import from schema
 import * as S from '@effect/schema/Schema'
+
 import { Email } from './email.ts'
 import { ISO8601Date } from './iso8601-date.ts'
 import { Phone } from './phone.ts'
