+++ apps/web/app/routes/resume/basics.tsx
@@ -4,9 +4,11 @@ import { SocialIcon } from '@suddenly-giovanni/ui/components/social/social-icon.
 import { T } from '@suddenly-giovanni/ui/components/typography/typography.tsx'
 import { clsx } from '@suddenly-giovanni/ui/lib/utils.ts'
 import type { ReactElement, ReactNode } from 'react'
+
 import { routesRecord } from '~/routes-record.ts'
-import type { BasicsType } from '~/schemas/server.resume/basics.ts'
-import type { ResumeType } from '~/schemas/server.resume/resume.ts'
+import type { ResumeType } from '~/.server/schemas/resume/resume.ts'
+import type { BasicsType } from '~/.server/schemas/resume/basics.ts'
+
 import resumePdfAssetUrl from '/giovanni-ravalico-resume-2021.pdf?url'
 
 export function Basics({ basics }: { basics: ResumeType['basics'] }): ReactElement {
