+++ apps/web/app/routes/resume/education.tsx
@@ -12,7 +12,9 @@ import * as Either from 'effect/Either'
 import { pipe } from 'effect/Function'
 import type { ReactElement } from 'react'
 import { memo, useCallback, useMemo, useState } from 'react'
-import type { EducationType } from '~/schemas/server.resume/education.ts'
+
+import type {EducationType} from "~/.server/schemas/resume/education.ts";
+
 import { formatDateLocaleShort } from './format-date-locale-short.ts'
 import { generateDjb2Hash } from './generate-djb2-hash.ts'
